Fixes bug in sessions# destroy method to only destroy an order_id in the sessions params if one exists.

+++ app/controllers/sessions_controller.rb
@@ -11,17 +11,19 @@ def create
   else
     flash.now[:error] = "Try again, password incorrect."
     render :new
-  end 
+  end
 end
 
 def destroy
   session[:merchant_id] = nil
 
+if session[:order_id]
   order = Order.find(session[:order_id])
   if order.status == "pending"
     order.destroy
   end
-  
+end
+
   redirect_to root_path
 end
